+++ src/app/_component/modal/_component/modalOutside/index.tsx
