+++ src/app/_component/modal/sign/MismatchPW/index.tsx
@@ -2,17 +2,14 @@ import ModalLayout from '@/src/app/_component/modal/_component/modalLayout';
 import ModalOutside from '@/src/app/_component/modal/_component/modalOutside';
 import ModalPortal from '@/src/app/_component/modal/_component/modalPortal';
 import ModalMainContent from '@/src/app/_component/modal/_component/modalMainContent';
-import { useSetRecoilState } from 'recoil';
-import { editPasswordState } from '@/src/app/_recoil/ModalAtom/sign';
 
-export default function Base({ content }: { content: string }) {
-  const setIsOpenUpdatedModal = useSetRecoilState(editPasswordState);
-  const handleClose = () => setIsOpenUpdatedModal(false);
+export default function MismatchPW() {
+  const handleClose = () => {};
   return (
     <ModalPortal>
       <ModalOutside closeModal={handleClose}>
         <ModalLayout btnName='확인' btnSize='large' sign={true} onClose={handleClose}>
-          <ModalMainContent content={content} />
+          <ModalMainContent content='비밀번호가 일치하지 않습니다.' />
         </ModalLayout>
       </ModalOutside>
     </ModalPortal>
