+++ src/app/_component/modal/sign/completeSignUp/index.tsx
@@ -3,11 +3,21 @@ import ModalOutside from '@/src/app/_component/modal/_component/modalOutside';
 import ModalPortal from '@/src/app/_component/modal/_component/modalPortal';
 import ModalMainContent from '@/src/app/_component/modal/_component/modalMainContent';
 import { useSetRecoilState } from 'recoil';
-import { editPasswordState } from '@/src/app/_recoil/ModalAtom/sign';
+import { completeSignUpState } from '@/src/app/_recoil/ModalAtom/sign';
+import { modalNameState } from '@/src/app/_recoil/ModalNameAtom';
+import { useEffect } from 'react';
 
-export default function Base({ content }: { content: string }) {
-  const setIsOpenUpdatedModal = useSetRecoilState(editPasswordState);
-  const handleClose = () => setIsOpenUpdatedModal(false);
+export default function CompleteSignUp() {
+  const content = '가입이 완료되었습니다!';
+  const setIsOpenSucessModal = useSetRecoilState(completeSignUpState);
+  const handleClose = () => setIsOpenSucessModal(false);
+  const setModalName = useSetRecoilState(modalNameState);
+
+  useEffect(() => {
+    setModalName(content);
+    return () => setModalName('');
+  }, []);
+
   return (
     <ModalPortal>
       <ModalOutside closeModal={handleClose}>
