--- src/app/_constant/modalType.tsx
@@ -1,38 +1,0 @@
-import { SubmitHandler, FieldValues } from 'react-hook-form';
-import { ToDoCardDetailProps } from '@/src/app/_component/modal/toDoCard/type';
-
-export const MODALTYPE = {
-  COLUMN: {
-    CREATE: '새 칼럼 생성',
-    UPDATE: '칼럼 관리',
-    DELETE: '칼럼 삭제',
-  },
-  TODO: {
-    CREATE: '할 일 생성',
-    UPDATE: '할 일 수정',
-    DELETE: '할 일 삭제',
-    DETAIL: '할 일 카드 상세',
-  },
-  DASHBOARD: {
-    CREATE: '새로운 대시보드',
-    INVITE: '초대하기',
-  },
-  SIGN: {
-    MISMATCH: '비밀번호가 일치하지 않습니다.',
-    COMPLETE: '가입이 완료되었습니다!',
-    ALREADY_USE: '이미 사용 중인 이메일입니다.',
-  },
-  MY_PAGE: {
-    WRONG_PW: '현재 비밀번호가 틀렸습니다.',
-  },
-};
-
-export interface CallModalType {
-  (condition: {
-    name: string;
-    onSubmit?: SubmitHandler<FieldValues>;
-    cardId?: number;
-    cardData?: ToDoCardDetailProps;
-    columnId?: number;
-  }): void;
-}
